Acomodar el date

+++ src/adoquines/AdoquinVariado.java
@@ -14,9 +14,9 @@ import adoquines.lista.Lista;
 public class AdoquinVariado {
     
     
-    public int  buscarCaminos(int n, int m){
+    public double  buscarCaminos(int n, int m){
         if(n<=1)return 1;
-        int i=0,cant=0;
+        double i=0,cant=0;
         do{
             cant+=buscarCaminos(--n,m);
             i++;
